Current float fix

+++ src/boot/ksdk1.1.0/devINA219.c
@@ -205,7 +205,7 @@ printSensorDataINA219(bool hexModeFlag, uint8_t address)
 		}
 		else if (address == 0x04)
 		{
-			Current = readSensorRegisterValueCombined / 10;
+			Current = (float)readSensorRegisterValueCombined / 10.0f;
 			warpPrint("Current : %d mA,\n", Current);
 		}
 		else
